test(#1828): share the rail's horizontal-inset reader, don't copy it
The #1802 gate grew a private reader for "which horizontal margin or
padding components does this rule body declare", complete with the
shorthand-arity expansion that makes `padding: 0.4rem 0` read as a zero
inline component rather than a 0.4rem one. The radio band needs exactly
the same question answered about a different region, and a second copy
of that arity table is a second place to get it wrong — the whole reason
the reader is subtle in the first place.

So it moves to `helpers/themeCss.ts`, next to `allRules` and
`selectorList` which it already depended on, and both gates import it.
`splitTopLevel` and `horizontalComponents` stay module-private there:
they are the implementation of the one exported question, not two more
questions.

Pure move, no behaviour change — `railInset.test.ts` is green on both
sides of it. `isZero` gains the `Inset` suffix on the way out, because a
bare `isZero` exported from a shared module says nothing about what it
is measuring.
